+++ weaver/processes/utils.py
@@ -1004,12 +1004,12 @@ def register_wps_processes_from_config(container, wps_processes_file_path=None):
     meaning they will be fetched on the provider each time a request refers to them, keeping their definition
     up-to-date with the remote server.
 
-    .. versionadded:: 1.14.0
+    .. versionadded:: 1.14
         When references are specified using ``providers`` section instead of ``processes``, the registration
         only saves the remote WPS provider endpoint to dynamically populate :term:`WPS` processes on demand.
         Previous behavior was to register each :term:`WPS` process individually with ID ``[service]_[process]``.
 
-    .. versionchanged:: 4.19.0
+    .. versionchanged:: 4.19
         Parameter position are inverted.
         If :paramref:`wps_processes_file_path` is explicitly provided, it is used directly without considering settings.
         Otherwise, automatically employ the definition in setting: ``weaver.wps_processes_file``.
@@ -1132,7 +1132,7 @@ def register_cwl_processes_from_config(container):
     """
     Load multiple :term:`CWL` definitions from a directory to register corresponding :term:`Process`.
 
-    .. versionadded:: 4.19.0
+    .. versionadded:: 4.19
 
     Each individual :term:`CWL` definition must fully describe a :term:`Process` by itself. Therefore, an ``id`` must
     be available in the file to indicate the target deployment reference. In case of conflict, the existing database